requirements.txt
python-dotenv==0.21.1
requests==2.28.2
streamlit==1.19.0
